% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstraps a sampling distribution for a sample.}
\usage{
bootstrap(sample, rep, n = "auto", estimator = mean, seed = NULL)
}
\arguments{
\item{sample}{A numeric vector to bootstrap}

\item{rep}{A numeric vector for number of replicates}

\item{n}{A numeric or character vector for the size of bootstrap samples}

\item{estimator}{A function that is one of \code{mean}, \code{median}, \code{var}, \code{sd}}

\item{seed}{A numeric vector as random seed. Can be \code{NULL}}
}
\value{
A numeric vector for bootstrap distribution
}
\description{
A sampling distribution of \code{rep} replicates is generated
for the specified \code{estimator} with replacement with a
bootstrap sample size of \code{n}.
}
\examples{
bootstrap(c(1, 2, 3), 3, 3)
bootstrap(c(1, 2, 3), 3, 3, estimator = var, seed = 1)
}
