% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_boot_stats.R
\name{calculate_boot_stats}
\alias{calculate_boot_stats}
\title{Calculates a bootstrapped confidence interval and other stats for a sample.}
\usage{
calculate_boot_stats(
  sample,
  rep,
  n = "auto",
  level = 0.95,
  estimator = "mean",
  seed = NULL,
  pass_dist = FALSE
)
}
\arguments{
\item{sample}{A numeric vector to bootstrap}

\item{rep}{A integer vector for number of replicates}

\item{n}{A integer or character vector for the size of bootstrap samples}

\item{level}{A numeric vector for the confidence level}

\item{estimator}{A character vector containing one of the
("mean", "median", "var", "sd") estimators}

\item{seed}{A integer vector as random seed. Can be \code{NULL}}

\item{pass_dist}{A boolean vector to decide whether to return the
bootstrapped distribution}
}
\value{
A named list
}
\description{
A bootstrapped confidence interval for the desired estimator for
the provided sample is calculated for a confidence level \code{level}.
Other stats and parameters of the distribution and sample are
also returned.
}
\examples{
calculate_boot_stats(c(1, 2, 3, 4), 1000, level = 0.95, seed = 1)
calculate_boot_stats(c(1, 2, 3, 4), 1000, level = 0.95, estimator = "median",
seed = 1, pass_dist = TRUE)
}
