% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Plots a bootstrapped sampling distribution with its confidence interval and
observed mean.}
\usage{
plot_ci(
  sample,
  rep,
  bin_size = 30,
  n = "auto",
  level = 0.95,
  random_seed = NULL,
  title = "",
  y_axis = "Count",
  estimator = "mean",
  path = NULL
)
}
\arguments{
\item{sample}{A numeric vector to bootstrap}

\item{rep}{A integer vector for number of replicates}

\item{bin_size}{A integer vector for number of bins representing intervals of
equal size over the range}

\item{n}{A integer or character vector for the size of bootstrap samples}

\item{level}{A numeric vector for the confidence level}

\item{random_seed}{A integer vector as random seed.}

\item{title}{A character vector for the title of the histogram}

\item{y_axis}{A character vector for the name of the y axis}

\item{estimator}{A character vector containing one of the("mean", "median",
"var", "sd") estimators}

\item{path}{A character vector for the path to directory from

current directory to save plot}
}
\value{
ggplot object
}
\description{
A bootstrapped confidence interval for the desired estimator for the provided
sample is calculated for a confidence level \code{level}.
}
\details{
Other stats and parameters of the distribution and sample are
also returned.
}
\examples{

plot_ci(c(1, 2, 3, 4, 5, 6, 7), 1000, n = 100, level = 0.95)
plot_ci(c(1, 2, 3, 4, 5, 6, 7), 1000, n = 100, path = "../")

}
