% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_stats.R
\name{tabulate_stats}
\alias{tabulate_stats}
\title{Makes two presentation-ready tables that summarize the statistics from the bootstrapped
samples and the parameters for creating the bootstrapped samples.}
\usage{
tabulate_stats(stat_list, precision = 2, path = NULL)
}
\arguments{
\item{stat_list}{A named list of the summary statistics produced
by the \code{calculate_boot_stats} function}

\item{precision}{A integer value for the precision of the
table values}

\item{path}{A character vector with the path to where the tables are
to be saved to}
}
\value{
a list containing 2 tibble objects:  table 1- summary statistics
table 2- bootstrapping parameters
}
\description{
A bootstrapped confidence interval for the desired estimator for
the provided sample is calculated for a confidence level \code{level}.
Other stats and parameters of the distribution and sample are
also returned.
}
\examples{
st <- calculate_boot_stats(c(1, 2, 3, 4), 1000, level = 0.95, seed = 123)
result  <-  tabulate_stats(st)
result[[1]] # stats table
result[[2]] # parameter table
}
