% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Plots a bootstrapped sampling distribution with its confidence
interval and observed mean.}
\usage{
plot_ci(
  sample,
  rep,
  bin_size = 30,
  n = "auto",
  ci_level = 0.95,
  ci_random_seed = NULL,
  title = "Bootstrap Sampling Plot",
  x_axis = "Bootstrap Sample Mean",
  y_axis = "Count",
  save_result_to = "."
)
}
\arguments{
\item{sample}{A numeric vector to bootstrap}

\item{rep}{A integer vector for number of replicates}

\item{bin_size}{A integer vector for number of bins representing
intervals of equal size over the range}

\item{n}{A integer or character vector for the size of bootstrap
samples}

\item{ci_level}{A numeric vector for the confidence level}

\item{ci_random_seed}{A integer vector as random seed.}

\item{title}{A character vector for the title of the histogram}

\item{x_axis}{A character vector for the name of the x axis}

\item{y_axis}{A character vector for the name of the y axis}

\item{save_result_to}{A character vector for the path to directory
from current directory to save plot}
}
\value{
ggplot object
}
\description{
A bootstrapped confidence interval for the desired estimator for
the provided sample is calculated for a confidence level \code{level}.
Other stats and parameters of the distribution and sample are
also returned.
}
\examples{
plot_ci(c(1, 2, 3, 4, 5, 6, 7), 1000, n = 100, 
                  ci_level = 0.95, ci_random_seed = 123)
plot_ci(c(1, 2, 3, 4, 5, 6, 7), 1000, n = 100, 
                 ci_level = 0.95, ci_random_seed = 123, 
                 save_result_to = "../")
}
